+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 ## [Unreleased](https://github.com/ixocreate/collection/compare/0.1.0...develop)
 ### Changed
 - Remove superfluous ServiceNotCreatedException
+- Performance update: only run strictUniqueKeys for generators
+- Performance update: use isset() instead of in_array for key check
+- Performance update: use Countable Interface (if available for input) for counting
 
 ## [v0.1.0 (2019-04-19)](https://github.com/ixocreate/collection/compare/master...0.1.0)
 ### Changed